Integrate consolidated completion extension
